requirements.txt
rpi.gpio==0.6.3
plantcv==3.1.0
